Make prototype of Serial_::begin consistent with implementation

+++ cores/arduino/USB/USBAPI.h
@@ -116,7 +116,7 @@ class Serial_ : public Stream
 public:
 	Serial_(USBDeviceClass &_usb) : usb(_usb), stalled(false) { }
 	void begin(uint32_t baud_count);
-	void begin(unsigned long, uint8_t);
+	void begin(uint32_t, uint8_t);
 	void end(void);
 
 	virtual int available(void);
